% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{autoplot.prcomp}
\alias{autoplot.prcomp}
\title{autoplot.prcomp}
\usage{
autoplot.prcomp(...)
}
\arguments{
\item{...}{Arguments passed to the \code{autoplot.prcomp} function.}
}
\value{
The result of the \code{autoplot.prcomp} function.
}
\description{
This function is a wrapper for the \code{autoplot.prcomp} function from the \code{ggfortify} package.
}
\keyword{re-export}
