% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyCompare-package.R
\docType{package}
\name{cyCompare-package}
\alias{cyCompare}
\alias{cyCompare-package}
\title{cyCompare: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggrlab/cyCompare}
  \item Report bugs at \url{https://github.com/ggrlab/cyCompare/issues}
}

}
\author{
\strong{Maintainer}: Gunther Glehr \email{gunthergl@gmx.net} (\href{https://orcid.org/0000-0002-1495-9162}{ORCID})

}
\keyword{internal}
