% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycompare_outcomes.R
\name{cycompare_outcomes}
\alias{cycompare_outcomes}
\title{Compare Flow Cytometry Data Across Devices}
\usage{
cycompare_outcomes(
  flowframes,
  df,
  ff_columns_relevant,
  dfcol_grouping_supersamples = c("Study"),
  dfcol_grouping_samples = "Device",
  dfcol_train_validation_other = "tvt",
  dfcol_outcomes = c("outcome_1", "outcome_2"),
  outcome_models = list(glmnet = glmnet::cv.glmnet),
  transformList = function(x) asinh(x/1000),
  gatingsets,
  gatename_primary,
  n_events_postgate = 10000,
  marker_to_gate,
  device_colors = function(n) {
     RColorBrewer::brewer.pal(n, "Dark2")
 },
  clustering_n_subsampling = 1,
  clustering_n_subsampled_cells = 10000,
  clustering_subsampling_seed_first = 42,
  clustering_outdir = tempdir(),
  kwargs_flowsom = list(nClus = 5, scale = FALSE, xdim = 3, ydim = 3, seed = 3711283),
  ...
)
}
\arguments{
\item{flowframes}{A named list of \code{flowFrame} objects containing flow cytometry data.}

\item{df}{A \code{data.table} containing metadata with at least the columns \code{"File"}, \code{"Device"}, and \code{"Sample"}.}

\item{ff_columns_relevant}{A character vector specifying the relevant markers for analysis.}

\item{gatingsets}{A named list of gating sets for each dataset.}

\item{gatename_primary}{A character string specifying the primary gating population.}

\item{n_events_postgate}{An integer specifying the maximum number of events to keep post-gating.}

\item{marker_to_gate}{A named vector mapping marker names to their corresponding gates.}

\item{device_colors}{A named vector or function that provides colors for each device.
If a function is provided, it should take the number of devices as input and return a vector of colors.}

\item{...}{Additional parameters passed to \code{plot_flowsom()}.}

\item{transformlist}{A transformation function or a named list of functions for transforming marker intensities
(default: \code{asinh(x / 1e3)}).}

\item{nClus}{An integer specifying the number of clusters for FlowSOM clustering (default: 5).}

\item{scale}{A logical indicating whether to scale the data in FlowSOM clustering (default: \code{FALSE}).}

\item{xdim}{An integer specifying the x-dimension of the FlowSOM grid (default: 3).}

\item{ydim}{An integer specifying the y-dimension of the FlowSOM grid (default: 3).}

\item{seed}{An integer specifying the random seed for FlowSOM clustering (default: \code{3711283}).}
}
\value{
A named list of ggplot2 objects containing:
\item{"Samples over time per device"}{A plot showing the number of samples collected over time per device.}
\item{"Counts and percentages"}{Plots of gated cell counts and percentages per sample.}
\item{"Positive population MFI"}{Plots showing the median fluorescence intensity (MFI) of positive gated populations.}
\item{"Density plots"}{Density distributions of marker intensities across devices and samples.}
\item{"Flowsom_PCA"}{Principal Component Analysis (PCA) plots of FlowSOM clustering results.}
\item{"Flowsom_MA"}{MA plots comparing cluster proportions between devices.}
}
\description{
This function performs a comparative analysis of flow cytometry data across multiple devices.
It includes basic sample statistics, gating, density plots, FlowSOM clustering, and marker intensity comparisons.
}
