% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsom_run_train.R
\name{flowsom_run_train}
\alias{flowsom_run_train}
\title{Train and Predict FlowSOM Clustering Based on Training Samples}
\usage{
flowsom_run_train(
  ff_gated,
  df,
  transformlist,
  seed = 1,
  dfcol_train_validation_other = NULL,
  relevant_columns,
  ...
)
}
\arguments{
\item{ff_gated}{A list of \code{flowFrame} objects or a \code{flowSet}, representing gated cytometry data.}

\item{df}{A \code{data.table} containing metadata with at least a \code{File} column. Only used
if \code{dfcol_train_validation_other} is specified, then \code{dfcol_train_validation_other} column
is used to select training samples for training the clustering.}

\item{transformlist}{Either a single transformation function or a named list of functions per marker.}

\item{seed}{Integer seed for reproducibility of subsampling.}

\item{dfcol_train_validation_other}{Optional column in \code{df} used to select training samples (e.g., \code{"TrainTest"}).}

\item{relevant_columns}{Character vector specifying markers to use for clustering. Defaults to all columns.}

\item{...}{Additional arguments passed to \code{flowsom_repeatsubsampling()}.}
}
\value{
A list containing cluster prediction results, as returned by \code{cytobench::flowSOM_predict()}.
}
\description{
This function applies marker transformation, runs FlowSOM clustering on a subset of training samples,
and predicts cluster assignments for all samples. Clustering is performed using \code{flowsom_repeatsubsampling()}.
}
