% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_grouped_apply.R
\name{fun_grouped_apply}
\alias{fun_grouped_apply}
\title{Apply a Function to Predefined Groupings of Flow Cytometry Data}
\usage{
fun_grouped_apply(
  data,
  result_grouping,
  make_flowset = TRUE,
  fun = flowsom_repeatsubsampling_apply,
  outdir_base = NULL,
  verbose = FALSE,
  return_results = TRUE,
  ...
)
}
\arguments{
\item{data}{The data which should be subset. E.g.
\itemize{
\item A named list of flowFrames, indexed by the "File" column of df.
}}

\item{result_grouping}{A list returned from \code{fun_grouped()} that contains \code{groups} and optionally \code{results}.}

\item{make_flowset}{Logical; if TRUE, convert \code{ff_list} into a \code{flowSet}. Default is TRUE.}

\item{fun}{A function to apply to each group. It should accept at least arguments \code{ff_list}, \code{outdir}, and a grouping-specific input (e.g., \code{result_grouping[["results"]][[i]]}).}

\item{outdir_base}{Optional string path to save results. Required if \code{return_results} is FALSE.}

\item{verbose}{Logical; if TRUE, prints progress messages.}

\item{return_results}{Logical; if FALSE, the function does not return results and assumes side effects (e.g., saving to disk).}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A list containing:
\describe{
\item{results}{List of results for each grouping, or NULL if \code{return_results} is FALSE.}
\item{groups}{Data frame of group combinations processed.}
}
}
\description{
This function applies a user-defined function (e.g., FlowSOM training) to data
across multiple groupings, typically the result of \code{fun_grouped()}. It allows optional conversion
to a \code{flowSet}, supports parallel execution, and can either return the results or rely on
side effects such as file output.
}
