% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{lossfun_hist_columnwise}
\alias{lossfun_hist_columnwise}
\title{Wrapper Functions for Histogram-Based Loss Calculation}
\usage{
lossfun_hist_columnwise(
  mat1,
  mat2,
  n_breaks = 100,
  lossfun = lossfun_hist_weighted,
  quantilebreaks = FALSE,
  result_percolumn = FALSE,
  ...
)
}
\arguments{
\item{mat1}{A numeric matrix. Each column represents a variable (e.g., marker) and rows are observations (e.g., cells).}

\item{mat2}{A second numeric matrix, same shape and interpretation as \code{mat1}.}

\item{n_breaks}{Integer (default: 100). Number of histogram bins to use. Must be at least 3.}

\item{lossfun}{A function that takes two matrices of binned counts and computes a loss. Default is \code{lossfun_hist_weighted}.}

\item{result_percolumn}{Logical (default: FALSE). If TRUE, returns a numeric vector of losses per column}

\item{...}{Additional arguments passed to \code{lossfun}.}
}
\description{
In contrast to \code{lossfun_hist()}, which computes the histogram-based loss
across all columns jointly, this function computes the loss column-wise and sums
the results.
}
\details{
NOTE: I am NOT sure if this is better or worse than \code{lossfun_hist()}.
}
