% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{lossfun_hist_cytonorm}
\alias{lossfun_hist_cytonorm}
\alias{lossfun_hist_weighted}
\title{Wrapper Functions for Histogram-Based Loss Calculation}
\usage{
lossfun_hist_cytonorm(x, y, ...)

lossfun_hist_weighted(x, y, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{emd_fun}.}
}
\value{
A numeric value representing the loss (EMD) between the input matrices.
}
\description{
These functions provide specific wrappers around \code{lossfun_hist()} for evaluating
differences between two matrices using Earth Mover's Distance (EMD). Both use
\code{emdist::emd2d()} with different argument configurations.
}
\details{
\code{\link[emdist]{emd2d}} interprets the two matrices A and B as a distibution
over a two-dimensional grid. The distance between the grid points in each direction
is defined by xdist and ydist.

\describe{
\item{\code{lossfun_hist_cytonorm}}{
Mimics the behavior of the CytoNorm implementation by discarding the \code{ydist}
argument before calling \code{emd2d()}.
}
\item{\code{lossfun_hist_weighted}}{
A generic weighted EMD computation that retains all arguments, including \code{ydist},
allowing full control over the distance calculation.
}
}
}
\examples{
mat1 <- matrix(rnorm(50), ncol = 10)
mat2 <- matrix(rnorm(50, mean = 1), ncol = 10)
lossfun_hist_cytonorm(mat1, mat2)
# Usually you would not use this function directly, but rather
# use lossfun_hist() which calls this internally.
lossfun_hist(mat1, mat2, n_breaks = 100, lossfun = lossfun_hist_cytonorm)
mat1 <- matrix(rnorm(50), ncol = 10)
mat2 <- matrix(rnorm(50, mean = 1), ncol = 10)
lossfun_hist_weighted(mat1, mat2)
# Usually you would not use this function directly, but rather
# use lossfun_hist() which calls this internally.
lossfun_hist(mat1, mat2, n_breaks = 100, lossfun = lossfun_hist_weighted)
}
\seealso{
\code{\link[emdist]{emd2d}}, \code{\link{lossfun_hist}}
}
