% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_grouped_apply.R
\name{models_grouped_apply}
\alias{models_grouped_apply}
\title{Train Grouped Models Using a Classification Pipeline}
\usage{
models_grouped_apply(
  df,
  outdir,
  result_grouping,
  grouping,
  result_grouping_models,
  dfcol_train_validation_other,
  dfcol_outcomes,
  bygroup = FALSE,
  ...
)
}
\arguments{
\item{df}{A data frame containing feature columns used for classification.}

\item{outdir}{A character string specifying the output directory to save model results. Not used here.}

\item{result_grouping}{A list object containing grouping results, including phenotype proportions.}

\item{grouping}{A data frame that defines grouping variables for merging with \code{df}.}

\item{result_grouping_models}{A list of models to be applied to the data.}

\item{dfcol_train_validation_other}{Character vector of column names indicating data split (e.g., train/validation/test).}

\item{dfcol_outcomes}{Character vector of outcome column names to be modeled.}

\item{bygroup}{Logical;
If FALSE, the function will apply each model to the entire dataset (result_grouping[\link{"clustering"}][\link{"proportions_per_x"}][\link{"metaCluster"}]).
If TRUE, the \code{FALSE} dataset is limited to the rows mathing the \code{grouping} variable.}

\item{...}{Not used.}
}
\value{
A list of trained models and evaluation results.
}
\description{
This function applies classification models to flow cytometry data grouped by metadata.
It handles preprocessing, joins with grouping information, and fits models using \code{cytobench::wrapper_count_models}.
}
