% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densities.R
\name{plot_densities}
\alias{plot_densities}
\title{Plot Density Distributions for Gated Flow Cytometry Data}
\usage{
plot_densities(
  ff_gated,
  df,
  device_colors,
  transformlist = NULL,
  density_n = 500,
  dfcol_grouping_samples = "Device",
  relevant_columns = NULL,
  limit_density_quantile = NA
)
}
\arguments{
\item{ff_gated}{A list of \code{flowFrame} objects containing gated flow cytometry data.}

\item{df}{A \code{data.table} containing metadata with at least the columns \code{"File"}, \code{"Device"}, and \code{"Sample"}.}

\item{device_colors}{An optional named vector of colors for each device to use in the plot.}

\item{transformlist}{A list of transformation functions for each marker. If a single function is provided, it will be applied to all markers.}

\item{density_n}{An integer specifying the number of points used to estimate density (default: 500).}

\item{relevant_columns}{A character vector of marker names to include in the density plots. If \code{NULL}, all markers will be included.}

\item{limit_density_quantile}{A numeric value between 0 and 1 to limit the density to a certain quantile within each variable.
If \code{NA}, no limit is applied and all densities are shown to their full extent. Use
if one device shows an extreme density peak that is not representative of the other devices
and thus deforms the density plot. 0.95 is a good value to start with.}
}
\value{
A \code{ggplot2} object showing the density distributions of markers across samples and devices.
}
\description{
This function generates density plots for multiple flow cytometry markers across different devices and samples.
}
