% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom_ma.R
\name{plot_flowsom_ma}
\alias{plot_flowsom_ma}
\title{Plot MA Plots Comparing Cluster Proportions Between Devices}
\usage{
plot_flowsom_ma(
  fs_pred,
  df,
  device_colors,
  dfcol_grouping_samples = "Device",
  dfcol_train_validation_other = NULL,
  MA_horizontal_lines_FC = c(2, 10, 25),
  MA_bins = 100
)
}
\arguments{
\item{fs_pred}{Output from \code{run_flowsom_clustering()} which comes from \code{cytobench::flowSOM_predict()} and
must contain the \code{ncells_per_x} list with the number of cells per cluster and sample.}

\item{df}{Metadata table.}

\item{device_colors}{Named color vector.}

\item{dfcol_grouping_samples}{Device column.}

\item{dfcol_train_validation_other}{Optional column for faceting.}

\item{MA_horizontal_lines_FC}{Fold-change reference lines.}

\item{MA_bins}{Number of bins for 2D histogram.}
}
\value{
A list of lists of MA plots.
}
\description{
Plot MA Plots Comparing Cluster Proportions Between Devices
}
