% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowsom_pca.R
\name{plot_flowsom_pca}
\alias{plot_flowsom_pca}
\title{Plot PCA of Cluster Proportions (FlowSOM)}
\usage{
plot_flowsom_pca(
  fs_pred,
  df,
  device_colors = NULL,
  dfcol_grouping_samples = "Device",
  dfcol_train_validation_other = NULL
)
}
\arguments{
\item{fs_pred}{Output from \code{run_flowsom_clustering()}.}

\item{df}{Metadata table.}

\item{device_colors}{Named color vector.}

\item{dfcol_grouping_samples}{Column used to color samples.}

\item{dfcol_train_validation_other}{Optional column used for shape/facets.}
}
\value{
A list of ggplot2 PCA plots.
}
\description{
Plot PCA of Cluster Proportions (FlowSOM)
}
