% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyCompare-package.R
\docType{package}
\name{cyCompare-package}
\alias{cyCompare}
\alias{cyCompare-package}
\title{cyCompare: Comparison and Evaluation of Flow Cytometry Normalization and Analysis Methods}
\description{
Provides tools to benchmark, compare, and visualize flow cytometry normalization and alignment methods across devices, locations, and time points. Implements a simplified optimal transport distance (OTD) to quantify device and sample variability, integrates FlowSOM clustering for population-level analysis, and supports Bland-Altman plots for device agreement. Designed for multi-center studies, with utilities for automated preprocessing and standardized output generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggrlab/cyCompare}
  \item Report bugs at \url{https://github.com/ggrlab/cyCompare/issues}
}

}
\author{
\strong{Maintainer}: Gunther Glehr \email{gunthergl@gmx.net} (\href{https://orcid.org/0000-0002-1495-9162}{ORCID})

}
\keyword{internal}
