% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsom_repeatsubsampling_apply.R
\name{flowsom_repeatsubsampling_apply}
\alias{flowsom_repeatsubsampling_apply}
\title{Apply FlowSOM prediction on repeated subsamplings}
\usage{
flowsom_repeatsubsampling_apply(
  ff_list,
  outdir,
  flowsom_result,
  n_metacluster = NULL,
  n_subsampling = 1,
  n_subsampled_cells = 10000,
  subsampling_seed_first = 427764,
  remove_results_keywords = c("flowsom_newdata", "cells_clusters_from_train"),
  ...
)
}
\arguments{
\item{ff_list}{A list of \code{flowFrame} objects to be subsampled and clustered.}

\item{outdir}{A character string indicating the output directory where results will be saved.
If \code{outdir} exists, the results will be written using \code{flowsom_wrapper_saving()}.}

\item{flowsom_result}{A trained FlowSOM object (typically returned by \code{flowsom_repeatsubsampling()} which called \code{cytobench::flowSOM_optimal()}).}

\item{n_metacluster}{Integer. Number of metaclusters to use when predicting clusters.
If NULL, the number of metaclusters in \code{flowsom_result} will be used.}

\item{n_subsampling}{Integer. Number of repeated subsampling iterations to perform. Default is 1.}

\item{n_subsampled_cells}{Integer. Number of cells per subsample. If the original sample is smaller,
upsampling is performed; otherwise, downsampling. If you want to use the original sample size,
set \code{n_subsampled_cells} to \code{Inf}. See \code{cytobench::subsample_ff()} for details.}

\item{subsampling_seed_first}{Integer. Base seed for reproducible subsampling. Each subsample iteration
uses \code{subsampling_seed_first + i - 1}.}

\item{remove_results_keywords}{Character vector. Names of result slots in \code{flowsom_result} to be removed from output.
This is useful to avoid returning large objects that are not needed for further analysis.}

\item{...}{Not used here, but potentially relevant if the calling function is used in a
pipeline with additional parameters.}
}
\value{
A nested list of FlowSOM prediction results. For each data slot, proportions of clusters per sample
are returned (after normalization). Slots defined in \code{remove_results_keywords} are removed. Each list entry
contains subsampling-specific results.
}
\description{
This function performs repeated subsampling of a list of flowFrames, applies a trained FlowSOM model
to the subsampled data, and returns predicted cluster results with normalized cluster proportions.
}
