% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_grouped_apply.R
\name{fun_grouped_apply}
\alias{fun_grouped_apply}
\title{Apply a Function to Predefined Groupings of Flow Cytometry Data}
\usage{
fun_grouped_apply(
  data,
  result_grouping,
  make_flowset = TRUE,
  fun = flowsom_repeatsubsampling_apply,
  outdir_base = NULL,
  verbose = FALSE,
  return_results = TRUE,
  ...
)
}
\arguments{
\item{data}{Named list of flowFrames or similar objects. Names must match \code{File} column from metadata.}

\item{result_grouping}{A list returned from \code{fun_grouped()}, containing \code{groups} and optionally \code{results}.
Can also be a list of such lists (e.g., clustering + modeling).}

\item{make_flowset}{Logical; if \code{TRUE}, convert the subset to a \code{flowSet} before calling \code{fun}.}

\item{fun}{Function to apply to each group. It must accept a subset of the data and an \code{outdir} as its first two arguments.}

\item{outdir_base}{Base directory for saving intermediate results (default: \code{tempdir()}).}

\item{verbose}{Logical; print progress messages (default: \code{FALSE}).}

\item{return_results}{Logical; whether to return the result of \code{fun()} for each group. If \code{FALSE},
assumes \code{fun()} is expected to save results to disk into \code{outdir_base}.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A named list with:
\describe{
\item{results}{List of outputs from \code{fun()} per group (or \code{NULL} if \code{return_results = FALSE}).}
\item{groups}{Data frame of group combinations used.}
}
}
\description{
Applies a user-defined function (e.g., FlowSOM application or modeling) to each group from \code{fun_grouped()} output.
This supports grouped application of downstream tasks across supersample/device combinations.
}
