% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_grouped.R
\name{models_grouped}
\alias{models_grouped}
\title{Train Grouped Models Using a Classification Pipeline}
\usage{
models_grouped(
  df,
  outdir,
  result_grouping,
  grouping,
  dfcol_train_validation_other,
  dfcol_outcomes,
  hparam_n_evaluations = 3,
  seed = 42,
  learners_classification = list(mlr3::lrn("classif.ranger", predict_type = "prob",
    predict_sets = c("train", "test"), max.depth = paradox::to_tune(2, 20), num.trees =
    paradox::to_tune(c(500, 1000, 1500, 2000)), importance = "impurity")),
  dv_class_positive = c(outcome_1 = "A", outcome_2 = 5.1),
  loss_measure = mlr3::msr("classif.logloss"),
  ...
)
}
\arguments{
\item{df}{A data frame containing feature columns used for classification.}

\item{outdir}{A character string specifying the output directory to save model results.}

\item{result_grouping}{A list object containing grouping results, including phenotype proportions.}

\item{grouping}{A data frame that defines grouping variables for merging with \code{df}.}

\item{dfcol_train_validation_other}{Character vector of column names indicating data split (e.g., train/validation/test).}

\item{dfcol_outcomes}{Character vector of outcome column names to be modeled.}

\item{hparam_n_evaluations}{Integer; number of hyperparameter evaluations during model tuning. Default is 3.}

\item{seed}{Integer; seed for reproducibility. Default is 42.}

\item{learners_classification}{A list of mlr3 learners with tuning parameters set. Default includes a tuned \code{classif.ranger}.}

\item{dv_class_positive}{A named vector specifying the positive class for each outcome variable.}

\item{loss_measure}{An \code{mlr3} measure object for model evaluation. Default is classification log-loss.}

\item{...}{Additional arguments passed to \code{cytobench::wrapper_count_models}.}
}
\value{
A list of trained models and evaluation results.
}
\description{
This function applies classification models to flow cytometry data grouped by metadata.
It handles preprocessing, joins with grouping information, and fits models using \code{cytobench::wrapper_count_models}.
}
